# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "nube api docs",
        "title": "GO Nube API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/devices/{uuid}": {
            "delete": {
                "description": "Delete Device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Delete Device",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "patch": {
                "description": "Update Device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Update Device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device",
                        "name": "Network",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/networks/{uuid}": {
            "get": {
                "description": "get by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "networks"
                ],
                "summary": "Show a Network",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Device ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Network"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update Network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "networks"
                ],
                "summary": "Update Network",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Network ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Network",
                        "name": "Network",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "description": "Delete Network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "networks"
                ],
                "summary": "Delete Network",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Network ID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/points/{id}": {
            "get": {
                "description": "get by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "points"
                ],
                "summary": "Show a Point",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Point ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "description": "Update Point",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "points"
                ],
                "summary": "Update Point",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Point ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Point",
                        "name": "Network",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "description": "Delete Point",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "points"
                ],
                "summary": "Delete Point",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Point ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Device": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "network_uuid": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Point"
                    }
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "model.Network": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Device"
                    }
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.Point": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "device_uuid": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "point_store": {
                    "$ref": "#/definitions/model.PointStore"
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "writeable": {
                    "type": "boolean"
                }
            }
        },
        "model.PointStore": {
            "type": "object",
            "properties": {
                "created_on": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "fault": {
                    "type": "boolean"
                },
                "fault_message": {
                    "type": "string"
                },
                "history_enable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "pointUuid": {
                    "type": "string"
                },
                "updated_on": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}